# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "ELv2",
      "identifier": "ELv2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "List all settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemSetting"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "update_setting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Setting updated successfully"
          },
          "400": {
            "description": "Invalid request or unauthorized key"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/settings/test": {
      "post": {
        "tags": [
          "crate::routes::settings"
        ],
        "operationId": "test_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connection test completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SystemSetting": {
        "type": "object",
        "description": "System Setting entry for Management Console",
        "required": [
          "key",
          "value",
          "category",
          "is_secret",
          "updated_at"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "is_secret": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "TestConnectionRequest": {
        "type": "object",
        "required": [
          "service",
          "url"
        ],
        "properties": {
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "service": {
            "type": "string"
          },
          "token": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "TestConnectionResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "UpdateSettingsRequest": {
        "type": "object",
        "required": [
          "key",
          "value",
          "category"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}